refactor: moved `CosmicWidgetSize` into render_implementations/widget_size.rs
